requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
google-auth>=2.15.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.70.0
selenium>=4.15.0
aiohttp>=3.8.0
langchain>=0.0.350
langchain-google-genai>=0.0.6
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
plotly>=5.17.0
webdriver-manager>=4.0.0